Use fixed point for du_dp in potentials (#556)

+++ timemachine/cpp/src/rmsd_restraint.cu
@@ -62,7 +62,7 @@ void RMSDRestraint<RealType>::execute_device(
     const double *d_box,
     const double lambda,
     unsigned long long *d_du_dx, // buffered
-    double *d_du_dp,
+    unsigned long long *d_du_dp,
     unsigned long long *d_du_dl, // buffered
     unsigned long long *d_u,     // buffered
     cudaStream_t stream) {
